Deconstruction of d3drender for 3d world and geometry
